use the project workspace dir as the working dir for deploy & publish
Makes it easier to build custom deploy and publish steps by
eliminating the requirement for the deploy/publish Docker images to
parse the Drone payload to determine the working dir (which virtually
all of them would need to do). With this change, deploy and publish
Docker images need not even be Drone-aware (they can just perform an
action in the current directory).