use context in Exec, respect cancelation
This allows the use of timeout and cancelable contexts in Exec.
Previously, it handles its own timeouts and signals by calling
os.Exit. This does not work when drone-exec is being used as a library
(because os.Exit exits the entire process). Using context enables
drone-exec to be used as a library.